Codemodded HasteModuleLoader to Runtime
