+++ src/Runtime/__tests__/NODE_PATH_dir/RegularModuleInNodePath.js
