+++ src/Runtime/__tests__/Runtime-NODE_PATH-test.js
@@ -15,8 +15,8 @@ jest.mock('../../environments/JSDOMEnvironment');
 const path = require('path');
 const normalizeConfig = require('../../config/normalize');
 
-describe('HasteModuleLoader', function() {
-  let HasteModuleLoader;
+describe('Runtime', function() {
+  let Runtime;
   let HasteResolver;
   let JSDOMEnvironment;
 
@@ -24,7 +24,7 @@ describe('HasteModuleLoader', function() {
   const rootPath = path.join(rootDir, 'root.js');
   const config = normalizeConfig({
     cacheDirectory: global.CACHE_DIRECTORY,
-    name: 'HasteModuleLoader-NODE_PATH-tests',
+    name: 'Runtime-NODE_PATH-tests',
     rootDir: path.resolve(__dirname, 'test_root'),
   });
 
@@ -33,14 +33,14 @@ describe('HasteModuleLoader', function() {
     const resolver = new HasteResolver(config, {resetCache: false});
     return resolver.getHasteMap().then(
       response => resolver.end().then(() =>
-        new HasteModuleLoader(config, environment, response)
+        new Runtime(config, environment, response)
       )
     );
   }
 
   function initHasteModuleLoader(nodePath) {
     process.env.NODE_PATH = nodePath;
-    HasteModuleLoader = require('../HasteModuleLoader');
+    Runtime = require('../Runtime');
     HasteResolver = require('../../resolvers/HasteResolver');
     JSDOMEnvironment = require('../../environments/JSDOMEnvironment');
   }
@@ -69,7 +69,7 @@ describe('HasteModuleLoader', function() {
 
   pit('doesnt find modules if NODE_PATH is relative', function() {
     const nodePath = process.cwd().substr(path.sep.length) +
-      'src/HasteModuleLoader/__tests__/NODE_PATH_dir';
+      'src/Runtime/__tests__/NODE_PATH_dir';
     initHasteModuleLoader(nodePath);
     return buildLoader().then(function(loader) {
       expect(() => {
