+++ src/Runtime/__tests__/test_root/GlobalImageStub.js
