+++ src/Runtime/__tests__/test_root/ModuleWithSideEffects.js
