+++ src/Runtime/__tests__/test_root/RelativeImageStub.js
