+++ src/Runtime/__tests__/test_root/throwing-fn.js
